Chatv3

+++ backend/src/friend/friend.controller.ts
@@ -17,13 +17,12 @@ export class FriendController {
     async makeFriend(@User() userId: string, @Param('pseudo') pseudo: string) {
         const user2: UserEntity = await this.userService.findByPseudo(pseudo);
         if (!user2)
-            throw new ErrorException(HttpStatus.NOT_FOUND, AboutErr.USER, TypeErr.NOT_FOUND);
+            throw new ErrorException(HttpStatus.NOT_FOUND, AboutErr.USER, TypeErr.NOT_FOUND, 'target not found');
         if (userId == user2.id)
-            throw new HttpException('cant add himself as friend !', HttpStatus.CONFLICT);
+            throw new ErrorException(HttpStatus.CONFLICT, AboutErr.USER, TypeErr.INVALID, 'cant add himself has friend');
         // si j'ai deja envoyer une requete a cet user ou bien si je suis deja ami avec lui
-        if (await this.userService.friendshipExist(userId, user2.id)) {
-            throw new HttpException('already friend(or in waiting)', HttpStatus.BAD_REQUEST);
-        }
+        if (await this.userService.friendshipExist(userId, user2.id))
+            throw new ErrorException(HttpStatus.BAD_REQUEST, AboutErr.USER, TypeErr.INVALID, 'already added friend');
         let eventName: string = 'newRequest';
         if (await this.userService.frienshipWaiting(userId, user2.id)) {
             await this.userService.acceptFriendship(userId, user2.id);
@@ -31,17 +30,18 @@ export class FriendController {
         }
         else
             await this.userService.createFriendship(userId, user2.id);
-        await this.chatGateway.friendEvent(eventName, user2.id, pseudo);
+        this.chatGateway.friendEvent(eventName, user2.id, pseudo);
     }
 
     @Delete('del/:pseudo')
     async deleteFriendship(@User() userId: string, @Param('pseudo') pseudo: string) {
         const user2: UserEntity = await this.userService.findByPseudo(pseudo);
         if (!user2)
-            throw new ErrorException(HttpStatus.NOT_FOUND, AboutErr.USER, TypeErr.NOT_FOUND);
-        
-            if (!await this.userService.friendshipExist(userId, user2.id))
-            throw new HttpException('friendship not exist', HttpStatus.BAD_REQUEST);
+            throw new ErrorException(HttpStatus.NOT_FOUND, AboutErr.USER, TypeErr.NOT_FOUND, 'target not found');
+        if (userId === user2.id)
+            throw new ErrorException(HttpStatus.BAD_REQUEST, AboutErr.USER, TypeErr.INVALID, 'invalid target(himself)');
+        if (!await this.userService.friendshipExist(userId, user2.id))
+            throw new ErrorException(HttpStatus.NOT_FOUND, AboutErr.USER, TypeErr.NOT_FOUND, 'friendship not exist');
         else
             return this.userService.removeFriendship(userId, user2.id);
     }
